requirements.txt
aiohttp==3.9.3
gradio==3.12.0
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
Pillow==10.2.0
PyYAML==6.0.1
python-dotenv==1.0.1
nest-asyncio==1.6.0
httpx==0.23.0
httpcore==0.15.0
